+++ src/Resources/app/administration/src/app/index.js
@@ -1,2 +1,4 @@
-import './components/heptacom-admin-open-auth-scope-field';
-import './components/heptacom-admin-open-auth-user-confirm-login';
+const { Component } = Shopware;
+
+Component.register('heptacom-admin-open-auth-scope-field', () => import('./components/heptacom-admin-open-auth-scope-field'));
+Component.register('heptacom-admin-open-auth-user-confirm-login', () => import('./components/heptacom-admin-open-auth-user-confirm-login'));