+++ src/Resources/app/administration/src/module/heptacom-admin-open-auth-client/page/create/index.js
@@ -1,9 +1,7 @@
 import './heptacom-admin-open-auth-client-create-page.scss';
 import template from './heptacom-admin-open-auth-client-create-page.html.twig';
 
-const { Component } = Shopware;
-
-Component.register('heptacom-admin-open-auth-client-create-page', {
+export default {
     template,
 
     inject: [
@@ -63,4 +61,4 @@ Component.register('heptacom-admin-open-auth-client-create-page', {
                 });
         }
     }
-});
+};