+++ src/Resources/app/administration/src/provider/open_id_connect/components/provider-settings/index.js
@@ -1,8 +1,6 @@
 import template from './provider-settings.html.twig';
 
-const { Component } = Shopware;
-
-Component.register('heptacom-admin-open-auth-provider-open-id-connect-settings', {
+export default {
     template,
 
     props: {
@@ -18,4 +16,4 @@ Component.register('heptacom-admin-open-auth-provider-open-id-connect-settings',
             }
         }
     }
-});
+};