+++ src/Resources/app/administration/src/provider/saml2/components/provider-settings/index.js
@@ -1,8 +1,6 @@
 import template from './provider-settings.html.twig';
 
-const { Component } = Shopware;
-
-Component.register('heptacom-admin-open-auth-provider-saml2-settings', {
+export default {
     template,
 
     props: {
@@ -71,4 +69,4 @@ Component.register('heptacom-admin-open-auth-provider-saml2-settings', {
             );
         },
     },
-});
+};